style/rearrange-files-and-colors
