+++ sparkpack/src/app/(main)/layout.tsx
@@ -1,0 +1,32 @@
+import type { Metadata } from "next";
+import { Geist, Geist_Mono } from "next/font/google";
+import "../globals.css"; 
+import TopNavbar from "@/constants/layout/TopNavbar";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist({ 
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
+
+export const metadata: Metadata = {
+  title: "Furrest Pet Insurance", 
+  description: "Developed by SparkPack",
+};
+
+export default function MainLayout({ 
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <TopNavbar />
+      {children}
+    </>
+  );
+}
