--- sparkpack/src/app/landing.tsx
@@ -1,2 +1,0 @@
-
-import { useState, useEffect } from "react";
