+++ sparkpack/src/app/layout.tsx
@@ -1,22 +1,5 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import TopNavbar from "@/constants/layout/TopNavbar";
+import "./globals.css"; // Your global styles
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
 
 export default function RootLayout({
   children,
@@ -25,11 +8,8 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        <TopNavbar />
-        {children}
+      <body>
+        {children} {/* This will render either (main) layout or auth pages directly */}
       </body>
     </html>
   );
